builtins: Add missing cast for __builtin_thread_pointer()
With --enable-builtins, we use __builtin_thread_pointer() for the
implementation of __get_tp(). However the former returns a void*,
whereas the return type for the latter is a uintptr_t.

Add an explicit cast to make the types match.
This fixes compilation with GCC.

Fixes: 9ad5f0fd84ec ("New option to choose compiler builtins")

+++ src/internal/pthread_impl.h
@@ -16,7 +16,7 @@
 #if defined(MUSL_USE_COMPILER_BUILTINS)
 static inline uintptr_t __get_tp()
 {
-	return __builtin_thread_pointer();
+	return (uintptr_t)__builtin_thread_pointer();
 }
 #endif
